requirements.txt
Gradio
langchain
openai
langchain-community
gradio
azure-cognitiveservices-vision-computervision
azure-ai-formrecognizer
azure-ai-textanalytics